problemstatement2/backenddjango/api/serializer.py
from api.models import User, AndroidApp, UserProfile, File
from django.contrib.auth.password_validation import validate_password
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework import serializers
from rest_framework.validators import UniqueValidator
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email')



class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)
        
        # These are claims, you can add custom claims
        token['name'] = user.userprofile.name
        token['username'] = user.username
        token['email'] = user.email
        token['points_earned'] = user.userprofile.points_earned
        token['image'] = str(user.userprofile.image)
        token['tasks_completed'] = user.userprofile.tasks_completed
        # ...
        return token


class RegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(
        write_only=True, required=True, validators=[validate_password])
    password2 = serializers.CharField(write_only=True, required=True)

    class Meta:
        model = User
        fields = ('email', 'username', 'password', 'password2')

    def validate(self, attrs):
        if attrs['password'] != attrs['password2']:
            raise serializers.ValidationError(
                {"password": "Password fields didn't match."})

        return attrs

    def create(self, validated_data):
        user = User.objects.create(
            username=validated_data['username'],
            email=validated_data['email']

        )

        user.set_password(validated_data['password'])
        user.save()

        return user

class AndroidAppSerializer(serializers.ModelSerializer):
    class Meta:
        model = AndroidApp
        fields = '__all__'

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ('id', 'user', 'name', 'image', 'points_earned', 'tasks_completed')
        read_only_fields = ('user',)  # Assuming user is not editable directly

    def validate_points_earned(self, value):
        if value < 0:
            raise serializers.ValidationError("Points earned cannot be negative")
        return value

    def validate_tasks_completed(self, value):
        if value < 0:
            raise serializers.ValidationError("Tasks completed cannot be negative")
        return value

class FileSerializer(serializers.ModelSerializer):
    class Meta:
        model = File
        fields = ('id', 'file')